# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/user_input_lines:
    get:
      tags:
        - UserInputLines
      summary: List UserInputLines
      description: Get a paginated list of UserInputLines with optional filtering and sorting
      operationId: listUserInputLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInputLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - UserInputLines
      summary: Create UserInputLine
      description: Create a new UserInputLine
      operationId: createUserInputLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInputLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInputLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    UserInputLine:
      type: object
      required:
        - id
        - user_input_id
        - survey_id
        - question_id
        - skipped
        - answered_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        user_input_id:
          type: string
          format: uuid
        survey_id:
          type: string
          format: uuid
        question_id:
          type: string
          format: uuid
        suggested_answer_id:
          nullable: true
          type: string
          format: uuid
        matrix_row_id:
          nullable: true
          type: string
          format: uuid
        skipped:
          type: boolean
        answer_type:
          nullable: true
          type: string
        value_char_box:
          nullable: true
          type: string
        value_text_box:
          nullable: true
          type: string
        value_numerical_box:
          nullable: true
          type: number
          format: float
        value_scale:
          nullable: true
          type: integer
          format: int32
        value_date:
          nullable: true
          type: string
          format: date
        value_datetime:
          nullable: true
          type: string
          format: date-time
        answer_score:
          nullable: true
          type: number
          format: float
        answer_is_correct:
          nullable: true
          type: boolean
        speed_seconds:
          nullable: true
          type: integer
          format: int32
        answered_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateUserInputLineRequest:
      type: object
      required:
        - user_input_id
        - survey_id
        - question_id
        - skipped
        - answered_at
        - metadata
      properties:
        user_input_id:
          type: string
          format: uuid
        survey_id:
          type: string
          format: uuid
        question_id:
          type: string
          format: uuid
        suggested_answer_id:
          nullable: true
          type: string
          format: uuid
        matrix_row_id:
          nullable: true
          type: string
          format: uuid
        skipped:
          type: boolean
        answer_type:
          nullable: true
          type: string
        value_char_box:
          nullable: true
          type: string
        value_text_box:
          nullable: true
          type: string
        value_numerical_box:
          nullable: true
          type: number
          format: float
        value_scale:
          nullable: true
          type: integer
          format: int32
        value_date:
          nullable: true
          type: string
          format: date
        value_datetime:
          nullable: true
          type: string
          format: date-time
        answer_score:
          nullable: true
          type: number
          format: float
        answer_is_correct:
          nullable: true
          type: boolean
        speed_seconds:
          nullable: true
          type: integer
          format: int32
        answered_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    UserInputLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserInputLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []